+++ src/test/groovy/com/github/jengelman/gradle/plugins/shadow/PublishingSpec.groovy
@@ -2,7 +2,9 @@ package com.github.jengelman.gradle.plugins.shadow
 
 import com.github.jengelman.gradle.plugins.shadow.util.AppendableMavenFileRepository
 import com.github.jengelman.gradle.plugins.shadow.util.PluginSpecification
+import org.apache.tools.zip.ZipFile
 import org.gradle.testkit.functional.ExecutionResult
+import spock.lang.Issue
 
 import java.util.jar.Attributes
 import java.util.jar.JarFile
@@ -155,76 +157,5 @@ class PublishingSpec extends PluginSpecification {
         assert dependency.groupId.text() == 'shadow'
         assert dependency.artifactId.text() == 'b'
         assert dependency.version.text() == '1.0'
-    }
-
-    def 'integration with application plugin'() {
-        given:
-        repo.module('shadow', 'a', '1.0')
-                .insertFile('a.properties', 'a')
-                .insertFile('a2.properties', 'a2')
-                .publish()
-
-        file('src/main/java/myapp/Main.java') << """
-            |package myapp;
-            |public class Main {
-            |   public static void main(String[] args) {
-            |       System.out.println("TestApp: Hello World! (" + args[0] + ")");
-            |   }
-            |}
-        """.stripMargin()
-
-        buildFile << """
-            |apply plugin: ${ShadowPlugin.name}
-            |apply plugin: 'application'
-            |apply plugin: 'java'
-            |
-            |mainClassName = 'myapp.Main'
-            |
-            |version = '1.0'
-            |
-            |repositories {
-            |   maven { url "${repo.uri}" }
-            |}
-            |
-            |dependencies {
-            |   compile 'shadow:a:1.0'
-            |}
-            |
-            |runShadow {
-            |   args 'foo'
-            |}
-        """.stripMargin()
-
-        settingsFile << "rootProject.name = 'myapp'"
-
-        when:
-        runner.arguments << 'runShadow'
-        runner.arguments << 'installShadow'
-        ExecutionResult result = runner.run()
-
-        then:
-        success(result)
-
-        and:
-        File installedJar = file('build/installShadow/myapp/lib/myapp-1.0-all.jar')
-        assert installedJar.exists()
-
-        and:
-        contains(installedJar, ['a.properties', 'a2.properties', 'myapp/Main.class'])
-
-        and:
-        JarFile jar = new JarFile(installedJar)
-        Attributes attributes = jar.manifest.mainAttributes
-        assert attributes.getValue('Main-Class') == 'myapp.Main'
-
-        then:
-        File startScript = file('build/installShadow/myapp/bin/myapp')
-        assert startScript.exists()
-
-        and:
-        assert startScript.text.contains("-jar \$APP_HOME/lib/myapp-1.0-all.jar")
-
-        and:
-        assert result.standardOutput.contains('TestApp: Hello World! (foo)')
     }
 }